+++ CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 ## [Unreleased](https://github.com/pack-it/binspect/compare/0.0.1...HEAD)
 
+### Added
+- Symlink traversal is now shown to the user.
+
 
 ## [v0.0.1](https://github.com/pack-it/binspect/releases/tag/0.0.1) - 2026-07-23